requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
langchain
langchain-google-genai
python-dotenv
